rename tests module
