+++ tests/conftest.py
@@ -9,7 +9,7 @@ from tucan.io import graph_from_file
 
 # Assertion rewriting
 # (see https://docs.pytest.org/en/stable/how-to/writing_plugins.html#assertion-rewriting)
-pytest.register_assert_rewrite("tucan.tests")
+pytest.register_assert_rewrite("tucan.test_utils")
 
 
 def graph_from_dimacs(filepath):
